Image_generator_creaai/requirements.txt
crewai
google-generativeai
langchain_google_genai
requests
pillow
python-dotenv
crewai-tools
python-dotenv